can't cast as int lookin ass

+++ tools4fools.lua
@@ -6,7 +6,7 @@ function getTime() --Return the turtle calculated time (custom unix time from ep
   hoursOfTicks = os.time() * 1000
   daysOfTicks = os.day() * 24 * 1000
   totalTicks = hoursOfTicks + daysOfTicks
-  return int(totalTicks/ticksPerSecond)
+  return totalTicks/ticksPerSecond
 end
  
 function setDirection(direction)
